# go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an Inventario daemon.",
        "title": "Inventario API",
        "contact": {
            "name": "Inventario Support",
            "url": "https://github.com/denisvmedia/inventario/issues",
            "email": "ask@artprima.cz"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/g/{groupSlug}/uploads/file": {
            "post": {
                "description": "Upload a single file of any type. The file is stored and a file entity is created without a linked entity.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/vnd.api+json"
                ],
                "tags": [
                    "uploads"
                ],
                "summary": "Upload a file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group slug",
                        "name": "groupSlug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "File to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/jsonapi.FileResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/jsonapi.Errors"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/g/{groupSlug}/uploads/restores": {
            "post": {
                "description": "Upload an XML backup file to be used for a restore operation.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/vnd.api+json"
                ],
                "tags": [
                    "uploads"
                ],
                "summary": "Upload restore file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group slug",
                        "name": "groupSlug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "XML backup file to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/jsonapi.UploadResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/jsonapi.Errors"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "jsonapi.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "user-level error message",
                    "type": "object"
                },
                "status": {
                    "description": "user-level status message",
                    "type": "string"
                }
            }
        },
        "jsonapi.Errors": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/jsonapi.Error"
                    }
                }
            }
        },
...
        "jsonapi.FileMeta": {
            "type": "object",
            "properties": {
                "signed_urls": {
                    "description": "Map of file ID to signed URLs and thumbnails",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/jsonapi.URLData"
                    }
                }
            }
        },
...
        "jsonapi.FileResponse": {
            "type": "object",
            "properties": {
                "attributes": {
                    "$ref": "#/definitions/models.FileEntity"
                },
                "id": {
                    "type": "string"
                },
                "meta": {
                    "description": "Optional meta field for signed URLs",
                    "allOf": [
                        {
                            "$ref": "#/definitions/jsonapi.FileMeta"
                        }
                    ]
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "files"
                    ],
                    "example": "files"
                }
            }
        },
...
        "jsonapi.URLData": {
            "type": "object",
            "properties": {
                "thumbnails": {
                    "description": "map of thumbnail size to signed URL",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "url": {
                    "description": "signed URL for file access",
                    "type": "string"
                }
            }
        },
...
        "jsonapi.UploadData": {
            "type": "object",
            "properties": {
                "fileNames": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "type": "string",
                    "example": "images"
                }
            }
        },
        "jsonapi.UploadResponse": {
            "type": "object",
            "properties": {
                "attributes": {
                    "$ref": "#/definitions/jsonapi.UploadData"
                },
                "id": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "uploads"
                    ],
                    "example": "uploads"
                }
            }
        },
...
        "models.FileCategory": {
            "type": "string",
            "enum": [
                "photos",
                "invoices",
                "documents",
                "other"
            ],
            "x-enum-varnames": [
                "FileCategoryPhotos",
                "FileCategoryInvoices",
                "FileCategoryDocuments",
                "FileCategoryOther"
            ]
        },
        "models.FileEntity": {
            "type": "object",
            "properties": {
                "category": {
                    "description": "Category is the user-meaningful classification surfaced in the UI\n(Photos/Invoices/Documents/Other).",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.FileCategory"
                        }
                    ]
                },
                "created_at": {
                    "description": "CreatedAt is when the file was created",
                    "type": "string"
                },
                "description": {
                    "description": "Description is an optional description of the file",
                    "type": "string"
                },
                "ext": {
                    "description": "Ext is the file extension including the dot.\nExample: \".pdf\"",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "linked_entity_id": {
                    "description": "LinkedEntityID is the ID of the linked entity (commodity or export)",
                    "type": "string"
                },
                "linked_entity_meta": {
                    "description": "LinkedEntityMeta contains metadata about the link type\nFor commodities: \"images\", \"invoices\", \"manuals\"\nFor exports: \"xml-1.0\" (version of the export file format)",
                    "type": "string"
                },
                "linked_entity_type": {
                    "description": "LinkedEntityType indicates what type of entity this file is linked to (commodity, export, or empty for standalone files)",
                    "type": "string"
                },
                "mime_type": {
                    "description": "MIMEType is the MIME type of the file.\nExample: \"application/pdf\"",
                    "type": "string"
                },
                "original_path": {
                    "description": "OriginalPath is the original filename as uploaded by the user.\nExample: \"invoice.pdf\"",
                    "type": "string"
                },
                "path": {
                    "description": "Path is the filename without extension. This is the only field that can be modified by the user.\nExample: \"invoice-2023\"",
                    "type": "string"
                },
                "tags": {
                    "description": "Tags are optional tags for categorization and search",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "description": "Title is the user-defined title for the file",
                    "type": "string"
                },
                "type": {
                    "description": "Type represents the category of the file (image, document, etc.)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.FileType"
                        }
                    ]
                },
                "updated_at": {
                    "description": "UpdatedAt is when the file was last updated",
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "models.FileType": {
            "type": "string",
            "enum": [
                "image",
                "document",
                "video",
                "audio",
                "archive",
                "other"
            ],
            "x-enum-varnames": [
                "FileTypeImage",
                "FileTypeDocument",
                "FileTypeVideo",
                "FileTypeAudio",
                "FileTypeArchive",
                "FileTypeOther"
            ]
        },
...
    }
}